Python test conducted on 04_16
Python test
Part 1 – Programming:
1. Write a program (with function) which takes a sequence of numbers and check if all 
numbers are unique.
2. Python program to find out the average of a set of integers
a. Note: Don’t use the pre-defined functions like np.mean
b. For ex: If input is 5, the output should be average of 1+2+3+4+5
3. Python program to check whether the given integer is a multiple of both 5 and 7
4. Python program to display the given integer in reverse manner
a. Ex: input 852, output should be 258
5. Create an inner function to calculate the addition in the following way
a. Create an outer function that will accept two parameters, a and b
b. Create an inner function inside an outer function that will calculate the addition of a 
and b
c. At last, an outer function will add 5 into addition and return it
6. Python Program to Check Leap Year
a. A leap year is exactly divisible by 4 except for century years (years ending with 00). 
The century year is a leap year only if it is perfectly divisible by 400
b. For ex: 2017 is not a leap year, 1900 is a not leap year, 2012 is a leap year, 2000 is a 
leap year
Part 2 – Pandas exercise (basic)
Note: Use automobile dataset for below questions
1. Find the most expensive car company name
2. Count total cars per company
3. Find each company’s highest price car
4. Sort all cars by Price column
5. Group by make, fuel-type and body-style to calculate the average of city-mpg and highwaympg (miles per gallon)
Part 3 - MCQ:
1 
2
3.
4.
5.
6.
7.
8.
9.
10.
11. 
12.
13.